pyproject.toml
[tool.poetry]
name = "graphql-training"
version = "0.1.0"
description = ""
authors = ["hoge"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.27"
strawberry-graphql = "^0.219.2"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
cryptography = "^42.0.4"
pymysql = "^1.1.0"
tortoise-orm = "^0.20.0"
aiomysql = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
